refactor - change mocks file structure
