+++ src/__mocks__/data/index.ts
@@ -1,0 +1,13 @@
+export { chatListMocks, messageInfoMocks } from './chat';
+export { friendMocks } from './friendMocks';
+export { hobbyData, terms } from './join';
+export { meetingBoardMocks } from './meetingBoardMocks';
+export { meetingDetailMocks } from './meetingDetailMocks';
+export { profileList } from './mutiProfile';
+export {
+  registerMeetingMocks,
+  requestListMocks,
+  requestMeetingMocks,
+} from './myMeeting';
+export { MOCK_URL } from './mypageMocks';
+export { noticeList } from './notice';
