+++ src/__mocks__/data/mypageMocks.ts
@@ -1,2 +1,4 @@
-export const MOCK_URL =
+const MOCK_URL =
   'https://img.freepik.com/free-photo/3d-rendering-zoom-call-avatar_23-2149556779.jpg?w=2000&t=st=1661578223~exp=1661578823~hmac=fcbafec28b2bc396b64f33424ed8fb434b852ba0e36ecddada0b49950562de78' as const;
+
+export { MOCK_URL };
